backend/requirements.txt
flask
flask_restful
flask_sqlalchemy
flask_script
flask_cors
flask_bootstrap
flask_wtf
gunicorn
requests
pymysql
uuid
flask_jwt_extended